% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DyadCensus.sna}
\alias{DyadCensus.sna}
\title{helper functions}
\usage{
DyadCensus.sna(i, data, sims, period, groupName, varName, levls = 1:3)
}
\arguments{
\item{cov}{numeric, covariate scores}
}
\description{
These functions calculate the respective network statistic for ego. When multiplied with the importance of each statistic (the 'parameters') this constitutes the network evaluation of ego. See: \code{\link[=ts_eval]{ts_eval()}}.
}
