% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{f_degree}
\alias{f_degree}
\alias{f_recip}
\alias{f_outAct}
\alias{f_inAct}
\alias{f_outPop}
\alias{f_inPop}
\alias{f_transTrip}
\alias{f_transMedTrip}
\title{Network Statistics}
\usage{
f_degree(net, ego)

f_recip(net, ego)

f_outAct(net, ego)

f_inAct(net, ego)

f_outPop(net, ego)

f_inPop(net, ego)

f_transTrip(net, ego)

f_transMedTrip(net, ego)
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{ego}{numeric, the ego for which we want to calculate the network statistic.}
}
\value{
numeric value
}
\description{
These functions calculate the respective network statistic for ego. When multiplied with the importance of each statistic (the 'parameters') this constitutues the network evaluation of ego. See: \code{\link[=f_eval]{f_eval()}}.
}
\details{
For examples on how to use these statistics see: \link{vignette("Introduction_RsienaTwoStep", package="RsienaTwoStep")}:.

For the mathematical definition of these network statistics see chapter 12 of the RSiena manual \insertCite{ripley2022manual}{RsienaTwoStep}.
}
\examples{
f_degree(net=net1, ego=3)

}
\references{
\insertRef{ripley2022manual}{RsienaTwoStep}
}
\seealso{
\code{\link[=f_eval]{f_eval()}}
}
\concept{networkstatistics}
