% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{f_eval}
\alias{f_eval}
\title{Evaluation function}
\usage{
f_eval(net, ego, statistics, parameters)
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{ego}{numeric, the ego for which we are going to calculat how its evaluates the network.}

\item{statistics}{list of names of statistic functions the same length as \code{parameters}. See \code{?f_degree} for a list of available statistic functions.}

\item{parameters}{vector of numeric values the same length as \code{statistics}.}
}
\description{
Evaluation function
}
\details{
For the mathematical definition of the evaluation function see chapter 12 of the RSiena manual \insertCite{ripley2022manual}{RsienaTwoStep}.
}
\examples{
f_eval(net=net1, ego=6, statistics=list(f_degree, f_recip),
parameters=c(-2,1))
f_eval(net=net1, ego=10, statistics=list(f_degree,
f_recip, f_transTrip, f_transMedTrip), parameters=c(-2,2,7,7))
}
\references{
\insertRef{ripley2022manual}{RsienaTwoStep}
}
\seealso{
\code{\link[=f_alternatives_ministep]{f_alternatives_ministep()}}, \code{f_alternatives_twostep}, \code{f_sims}
}
