% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_rsiena.R
\name{f_select}
\alias{f_select}
\title{Select two actors}
\usage{
f_select(net, steps = 1)
}
\arguments{
\item{steps}{numeric, do we want to sample one or two actors}

\item{name}{matrix, the adjacency matrix representing the relations between actors}
}
\value{
vector of length one or two, with the actors being sampled.
}
\description{
Select two actors
}
\details{
This function selects two actors at random that are part of the network
}
\examples{
f_select(net=matrix(1:100, nrow=10, ncol=10), steps=2)

}
