% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sims.R
\name{f_sims}
\alias{f_sims}
\title{Simulation of Network evolution}
\usage{
f_sims(
  nsims = 1000,
  parallel = FALSE,
  net,
  rate,
  statistics,
  parameters,
  p2step = 0,
  chain = FALSE,
  dist1 = NULL,
  dist2 = NULL,
  modet1 = "degree",
  modet2 = "degree"
)
}
\arguments{
\item{nsims}{numeric, number of simulations.}

\item{parallel}{TRUE/FALSE}

\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{rate}{numeric, the average number of possible tie-changes per actor in the simulation.}

\item{statistics, }{list of names of statistic functions (see e.g. \code{\link[=f_degree]{f_degree()}} for a list of available functions)}

\item{parameters, }{vector of numeric values the same length as \code{parameters}}

\item{p2step}{numeric, value between range \verb{[0,1]} setting the probability that a twostep will occur relative to a ministep.}

\item{chain}{TRUE/FALSE, set to \code{TRUE} if you want to save all the subsequent networks (after the ministep or twostep) during the simulation. If \code{FALSE} only the end network is saved.}

\item{dist1}{numeric, minimal path length between ego1 and ego2 at time1 in order to be allowed to start a coorporation. If \code{NULL} all dyads are allowed to start a cooperation.}

\item{dist2}{numeric, minimal path length between ego1 and ego2 at time2 in order for twostep to be counted as coorporation.}

\item{modet1}{string, indicating the type of ties being evaluated at time1. "\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows directed paths starting from ego1 and ending at ego2. "\code{indegree}" only allows directed paths starting from ego2 and ending at ego2.}

\item{modet2}{string, indicating the type of ties being evaluated at time2. "\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows directed paths starting from ego1 and ending at ego2. "\code{indegree}" only allows directed paths starting from ego2 and ending at ego2.}
}
\value{
If \code{chain=FALSE} a \code{list} (of length \code{nsims}) of adjacency matrices representing the final network after the simulated evolution.
If \code{chain=TRUE} a \code{list} of lists of adjacency matrices. Each inner list represents the complete network evolution of one simulation. The outer list refers to the simulation run (with length \code{nsims}).
}
\description{
\code{f_sims} is the workhorse function of the package \code{RsienaTwoStep}.
It simulates the network evolution \code{nsims} times given an existing network \code{net},
the defined evaluation function (\code{\link[=f_eval]{f_eval()}}) by \code{statistics} and \code{parameters} and
the average number of possible tie changes per actor as defined by \code{rate}.
}
\details{
For examples on how to use \code{f_sims} see: \link{vignette("Introduction_RsienaTwoStep", package="RsienaTwoStep")}: or the \href{https://jochemtolsma.github.io/RsienaTwoStep/}{package website}.
Before you set \code{parallel} to TRUE make sure to set-up a cluster with the package \code{doParallel} (see \code{Examples}).
}
\examples{
\dontrun{#simulation with ministep only
f_sims(net=net1, rate=5, statistics=list(f_degree, f_recip),
parameters=c(-3,1))

#simulation with twosteps only f_sims(net=net1, rate=5,
statistics=list(f_degree, f_recip), parameters=c(-3,1), p2step=1)

#running simulations in parallel n.cores <- parallel::detectCores() - 1
#save one core for other work #create the cluster my.cluster <-
parallel::makeCluster(n.cores, type = "PSOCK") #register it to be used by
\%dopar\% #doParallel::registerDoParallel(cl = my.cluster) f_sims(net=net1,
rate=5, parallel = TRUE, statistics=list(f_degree, f_recip),
parameters=c(-3,1), p2step=1) }
}
\seealso{
\code{\link[=f_alternatives_ministep]{f_alternatives_ministep()}}, \code{\link[=f_alternatives_twostep]{f_alternatives_twostep()}}, \code{\link[=f_eval]{f_eval()}}
}
