% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net1.R
\docType{data}
\name{net1}
\alias{net1}
\title{net1}
\format{
\subsection{\code{net1}}{

An non-symmetric (adjacency) matrix of 10 rows and 10 columns: 0 indicates no tie; 1 indicates directed tie (from ego (row) to alter (column)).
}
}
\usage{
net1
}
\description{
A fictive directed network
}
\keyword{datasets}
