% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives_beh.R
\name{ts_alternatives_ministep_beh}
\alias{ts_alternatives_ministep_beh}
\alias{ts_alternatives_simstep_beh}
\alias{ts_alternatives_twostep_beh}
\title{Possible behavior after ministep, simstep or twostep}
\usage{
ts_alternatives_ministep_beh(beh, ego)

ts_alternatives_simstep_beh(beh, ego)

ts_alternatives_twostep_beh(beh, net, dist1 = NULL, modet1 = "degree")
}
\arguments{
\item{beh}{numerical, vector representing the behavioral scores
actors.}

\item{ego}{numeric, value indicating ego (row number of net)}

\item{net}{matrix, the adjacency matrix representing the relations between
actors. Valid values are 0 and 1.}

\item{dist1}{numeric, minimal path length between ego1 and ego2 at time1 in
order to be allowed to start a coordination. If \code{NULL} all dyads are
allowed to start a coordination (i.e. \strong{simultaneity}).}

\item{modet1}{string indicating the type of ties being evaluated at time1.
"\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows
directed paths starting from ego1 and ending at ego2. "\code{indegree}" only
allows directed paths starting from ego2 and ending at ego1. See:
\code{DETAILS}.}
}
\value{
list, a list of the alternative vector representing te behavioral
scores
}
\description{
\code{\link[=ts_alternatives_ministep_beh]{ts_alternatives_ministep_beh()}} constructs the possible future
behavior score after a ministep of \code{ego}.
\code{\link[=ts_alternatives_twostep_beh]{ts_alternatives_twostep_beh()}} constructs the possible future behavioral
scores after two ministeps of two egos.
\code{\link[=ts_alternatives_simstep_beh]{ts_alternatives_simstep_beh()}} constructs all possible future behavioral
scores (over the complete range of the behavioral variable) of one ego.Ego is
thus allowed to jump from one extreme to the other
}
\examples{
ccovar <- ts_prepdata(df_ccovar1)
ts_alternatives_ministep_beh(beh = ccovar[, "cov2"], ego = 3)
}
\seealso{
\code{\link[=ts_select]{ts_select()}}, \code{\link[=ts_sims]{ts_sims()}}
}
