% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ts_degree}
\alias{ts_degree}
\alias{ts_recip}
\alias{ts_outAct}
\alias{ts_inAct}
\alias{ts_outPop}
\alias{ts_inPop}
\alias{ts_transTrip}
\alias{ts_transMedTrip}
\alias{ts_transRecTrip}
\alias{ts_cycle3}
\title{Network Statistics}
\usage{
ts_degree(net, ego)

ts_recip(net, ego)

ts_outAct(net, ego)

ts_inAct(net, ego)

ts_outPop(net, ego)

ts_inPop(net, ego)

ts_transTrip(net, ego)

ts_transMedTrip(net, ego)

ts_transRecTrip(net, ego)

ts_cycle3(net, ego)
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{ego}{numeric, the ego for which we want to calculate the network statistic.}
}
\value{
numeric value
}
\description{
These functions calculate the respective network statistic for ego. When multiplied with the importance of each statistic (the 'parameters') this constitutues the network evaluation of ego. See: \code{\link[=ts_eval]{ts_eval()}}.
}
\details{
For examples on how to use these statistics see: \code{vignette("1.Introduction_RsienaTwoStep", package="RsienaTwoStep")}.

For the mathematical definition of these network statistics see chapter 12 of the RSiena manual \insertCite{ripley2022manual}{RsienaTwoStep}.
}
\examples{
ts_degree(net=net1, ego=3)

}
\references{
\insertRef{ripley2022manual}{RsienaTwoStep}
}
\seealso{
\code{\link[=ts_eval]{ts_eval()}}
}
\concept{networkstatistics}
