% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{ts_geodist}
\alias{ts_geodist}
\title{Distance between alters}
\usage{
ts_geodist(net, ego1, ego2, degree = "degree")
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{ego1}{numeric, value indicating ego1 (row number of net)}

\item{ego2}{numeric, value indicating ego2 (row number of net)}

\item{degree}{string, the type of path to be considered: "\code{outdegree}"; "\code{indegree}"; "\code{degree}".}
}
\value{
numeric vector of length one, with the distance between the two egos. If egos are not connected \code{ts_geodist} returns \code{Inf}.
}
\description{
Calculates the distance, or path length, between two actors part of the same network.
}
\details{
This function is a wrapper around the function \code{\link[sna:geodist]{sna::geodist}}.
}
\examples{
ts_geodist(net=net1, ego1=6, ego2=8, degree="outdegree")
ts_geodist(net=net1, ego1=8, ego2=6, degree="outdegree")
ts_geodist(net=net1, ego1=8, ego2=6, degree="degree")
}
