% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_beh.R
\name{ts_linear}
\alias{ts_linear}
\alias{ts_quad}
\alias{ts_avAlt}
\alias{ts_effFrom}
\title{Behavior Statistics}
\usage{
ts_linear(beh, ego)

ts_quad(beh, ego)

ts_avAlt(beh, net, ego, cov = NULL)

ts_effFrom(beh, net = NULL, ego, cov)
}
\arguments{
\item{beh}{behavioral dependent variable}

\item{ego}{numeric, the ego for which we want to calculate the network
statistic.}

\item{net}{matrix, the adjacency matrix representing the relations between
actors. Valid values are 0 and 1.}

\item{cov}{numeric, covariate scores}
}
\value{
numeric value
}
\description{
These functions calculate the respective behavior statistic for
ego. When multiplied with the importance of each statistic (the
'parameters') this constitutes the network evaluation of ego. See:
\code{\link[=ts_eval]{ts_eval()}}.
}
\details{
For examples on how to use these statistics see:
\code{vignette("Introduction_RsienaTwoStep", package="RsienaTwoStep")}.

For the mathematical definition of these network statistics see chapter 12
of the RSiena manual \insertCite{ripley2022manual}{RsienaTwoStep}.
}
\examples{
ts_linear(df_ccovar1$cov2, ego=3)

}
\references{
\insertRef{ripley2022manual}{RsienaTwoStep}
}
\seealso{
\code{\link[=ts_eval]{ts_eval()}}

Other networkstatistics: 
\code{\link{ts_degree}()}
}
\concept{networkstatistics}
