% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net1.R
\docType{data}
\name{ts_net1}
\alias{ts_net1}
\title{ts_net1}
\format{
\subsection{\code{ts_net1}}{

An non-symmetric (adjacency) matrix of 10 rows and 10 columns: 0 indicates no tie; 1 indicates directed tie (from ego (row) to alter (column)).
}
}
\usage{
ts_net1
}
\description{
A fictive directed network
}
\keyword{datasets}
