% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net2.R
\docType{data}
\name{ts_net2}
\alias{ts_net2}
\title{ts_net2}
\format{
\subsection{\code{ts_net2}}{

An non-symmetric (adjacency) matrix of 5 rows and 5 columns: 0 indicates no tie; 1 indicates directed tie (from ego (row) to alter (column)).
}
}
\usage{
ts_net2
}
\description{
A fictive directed network
}
\keyword{datasets}
