% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_simdata.R
\docType{data}
\name{ts_simdata}
\alias{ts_simdata}
\title{ts_simdata}
\format{
\subsection{\code{ts_simdata}}{

A list containing two adjacency networks (net_sim1 and net_sim2), a fictive
dataset (DF_sim) and information on the simulation process
(startvalues_sim, statistics_sim, p2_step_sim, dist1_sim).
}
}
\usage{
ts_simdata
}
\description{
A simulated co-evoluation of friendships and alcohol consumption
}
\keyword{datasets}
