% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_alternatives_ministep.R
\name{f_alternatives_ministep}
\alias{f_alternatives_ministep}
\alias{f_alternatives_twostep}
\title{Possible networks after ministep of ego}
\usage{
f_alternatives_ministep(net, ego)

f_alternatives_twostep(
  net,
  dist1 = NULL,
  dist2 = NULL,
  modet1 = "degree",
  modet2 = "degree"
)
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{ego}{numeric, value indicating ego (row number of net)}

\item{dist1}{numeric, minimal path length between ego1 and ego2 at time1 in order to be allowed to start a coorporation. If \code{NULL} all dyads are allowed to start a cooperation.}

\item{dist2}{numeric, minimal path length between ego1 and ego2 at time2 in order for twostep to be counted as coorporation. See \code{DETAILS}.}

\item{mode1}{string indicating the type of ties being evaluated at time1. "\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows directed paths starting from ego1 and ending at ego2. "\code{indegree}" only allows directed paths starting from ego2 and ending at ego2. See: \code{DETAILS}.}

\item{mode2}{string, indicating the type of ties being evaluated at time2. "\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows directed paths starting from ego1 and ending at ego2. "\code{indegree}" only allows directed paths starting from ego2 and ending at ego2. See: \code{DETAILS}.}
}
\value{
list, a list of the alternative adjacency matrices after all possible ministeps of ego (\code{f_alternatives_ministep}) or after all possible twosteps of two egos (\code{f_alternatives_twostep})
}
\description{
\code{f_alternatives_ministep} constructs the possible future networks at time2 after a ministep of \code{ego} given the network \code{net} at time1.
\code{f_alternatives_twostep} constructs the possible future networks at time2 after a twostep of two internally sampled egos (via \code{\link{f_select}}) given the network \code{net} at time1.
}
\details{
\code{f_alternatives_ministep} mimics the ministep assumption as implemented in the SAOM of \code{\link{RSiena}} \insertCite{ripley2022manual}{RsienaTwoStep}.
\code{f_alternatives_twostep} allows two actors to simultaneously make a ministep, that is a \strong{twostep}.
Further restrictions can be set to which actors are allowed to make a twostep:
\enumerate{
\item Two random actors
\item Two actors who are connected at time1 in a specific way (determined by \code{dist1} and \code{mode1})
}

A special case is to allow two random actors to make a twostep but only consider some of the alternative future networks as the result of cooperation, namely those networks that:
\enumerate{
\item result from actors who are connected at time1 in a specific way (determined by \code{dist1} and \code{mode1}), \strong{or}
\item where actors are connected at time2 in a specific way (determined by \code{dist2} and \code{mode2})
}
}
\examples{
f_alternatives(net=net1, ego=3)
}
\references{
\insertRef{ripley2022manual}{RsienaTwoStep}
}
\seealso{
f_select
}
