% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_select.R
\name{f_select}
\alias{f_select}
\title{Select actor(s)}
\usage{
f_select(net, steps = 1, dist1 = NULL, modet1 = "degree")
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{steps}{numeric, do we want to sample one or two actors}

\item{dist1}{numeric, minimal path length between ego1 and ego2 at time1 in order to be allowed to start a coordination. If \code{NULL} just simultaneity. If not \code{NULL} \code{steps} is set to 2.}

\item{modet1}{string indicating the type of ties being evaluated at time1. "\code{degree}" considers all ties as undirected. "\code{outdegree}" only allows directed paths starting from ego1 and ending at ego2. "\code{indegree}" only allows directed paths starting from ego2 and ending at ego2.}
}
\value{
vector of length one or two, with the actors being sampled.
}
\description{
Select one or two actors that will perform a (simultaneous) ministep.
}
\details{
This function selects one actor (\code{steps=1}) or two actors (\code{steps=2}) at random that are part of the network. If it is not possible to select two actors to start a coordination (not connected dyads at time1), the function will tell you so.
}
\examples{
f_select(net=net1, steps=1)
f_select(net=net1, steps=2)
f_select(net=net1, dist1=2, modet1="degree")

}
