% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_select.R
\name{f_select}
\alias{f_select}
\title{Select two actors}
\usage{
f_select(net, steps = 1)
}
\arguments{
\item{net}{matrix, the adjacency matrix representing the relations between actors. Valid values are 0 and 1.}

\item{steps}{numeric, do we want to sample one or two actors}
}
\value{
vector of length one or two, with the actors being sampled.
}
\description{
Select two actors
}
\details{
This function selects two actors at random that are part of the network
}
\examples{
f_select(net=net1, steps=2)

}
