% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{ts_centering}
\alias{ts_centering}
\alias{ts_simij}
\title{Data preparation}
\usage{
ts_centering(cov)

ts_simij(cov, min = NULL, max = NULL)
}
\arguments{
\item{cov}{numeric, behavioral scores of actors}

\item{min}{numeric, minimum value of behavioral scores of actors. If \code{NULL} the empirically observed minimum is used.}

\item{max}{numeric, maximum value of behavioral scores of actors. If \code{NULL} the empirically observed maximum is used.}
}
\description{
\code{ts_centering} centers the variables before use.
\code{ts_simij} calculates the similarity scores before use.
}
\details{
to do
}
\examples{
ts_centering(cov=df_ccovar1[,"cov1"])
ts_simij(cov=df_ccovar1[,"cov2"])
ts_simij(cov=df_ccovar1[,"cov2"], min=-5, max=5)
}
