% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ts_centering}
\alias{ts_centering}
\title{helper functions}
\usage{
ts_centering(cov)
}
\arguments{
\item{cov}{numeric, covariate scores}
}
\description{
These functions calculate the respective network statistic for ego. When multiplied with the importance of each statistic (the 'parameters') this constitutes the network evaluation of ego. See: \code{\link[=ts_eval]{ts_eval()}}.
}
